refactor(llmz): split vm.ts into modular driver architecture
Break the 1286-line vm.ts monolith into focused modules:

  src/vm/
    index.ts          — runAsyncFunction dispatcher
    types.ts          — VMDriver interface
    instrument.ts     — shared instrumentation (both drivers)
    errors.ts         — error formatting for both drivers
    drivers/
      quickjs.ts      — sandboxed QuickJS WASM driver
      node.ts         — unsandboxed Node fallback driver

No behavioral changes — all 382 tests pass unchanged.
